requirements.txt
aiohttp==3.6.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
discord.py==1.5.1
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
Naked==0.1.31
pycparser==2.20
pylint==2.6.0
PyYAML==5.3.1
requests==2.25.1
shellescape==3.8.1
six==1.15.0
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
yarl==1.5.1
